+++ TablePro/Views/ConnectionForm/ConnectionFormWindowController.swift
@@ -68,7 +68,9 @@ internal final class ConnectionFormWindowController: NSWindowController, NSWindo
         window.isRestorable = false
         window.contentMinSize = NSSize(width: 720, height: 560)
         window.setContentSize(NSSize(width: 820, height: 620))
-        window.setFrameAutosaveName(NSWindow.FrameAutosaveName(SplitViewAutosaveName.current(WindowIdentifier.connectionForm)))
+        if let autosaveName = SplitViewAutosaveName.current(WindowIdentifier.connectionForm) {
+            window.setFrameAutosaveName(autosaveName)
+        }
 
         self.init(window: window)